renamed backward_compatibility
