+++ test/backward_compatibility/shared/src/worker.go
